feat(jsonutil): 添加自定义JSON处理工具包
refactor: 将int64类型字段统一改为jsonutil.Int64以支持字符串序列化
refactor(handlers): 更新post和category处理逻辑以使用jsonutil.Int64
refactor(models): 修改DTO结构体使用jsonutil.Int64替代原生int64
refactor(session): 更新session存储结构使用jsonutil.Int64
refactor(storage): 调整minio客户端返回类型使用jsonutil.Int64
test(jsonutil): 添加Int64类型的序列化测试用例

+++ backend/handlers/post.go
@@ -36,7 +36,7 @@ func (h *PostHandler) Create(c *gin.Context) {
 	clientIP := middleware.ExtractClientIP(c.Request.RemoteAddr)
 	userAgent := c.Request.UserAgent()
 
-	post, err := h.postService.CreatePost(c.Request.Context(), user.ID, &req, clientIP, userAgent)
+	post, err := h.postService.CreatePost(c.Request.Context(), int64(user.ID), &req, clientIP, userAgent)
 	if err != nil {
 		switch {
 		case errors.Is(err, service.ErrPostCooldown):
@@ -112,7 +112,7 @@ func (h *PostHandler) Update(c *gin.Context) {
 		return
 	}
 
-	post, err := h.postService.UpdatePost(c.Request.Context(), id, user.ID, &req)
+	post, err := h.postService.UpdatePost(c.Request.Context(), id, int64(user.ID), &req)
 	if err != nil {
 		switch {
 		case errors.Is(err, service.ErrPostNotFound):
@@ -147,7 +147,7 @@ func (h *PostHandler) Delete(c *gin.Context) {
 
 	isAdmin := false
 
-	err = h.postService.DeletePost(c.Request.Context(), id, user.ID, isAdmin, req.Reason)
+	err = h.postService.DeletePost(c.Request.Context(), id, int64(user.ID), isAdmin, req.Reason)
 	if err != nil {
 		switch {
 		case errors.Is(err, service.ErrPostNotFound):
